requirements.txt
fastapi
uvicorn
SQLAlchemy
pydantic
passlib[bcrypt]
python-jose[cryptography]
